requirements.txt
Pillow
einops
safetensors
timm
tomesd
torchdiffeq
torchsde
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
onnxruntime
accelerate>=0.25.0
gradio>=3.41.2
diffusers>=0.30.1
transformers>=4.46.2
deepspeed
jaxtyping
hydra-core
torch_scatter